Add files via upload
Here we gonna learn how to loop backwards in array. It means that we keep on printing the elements of the array starting from the last one to the first one. We can also be able to insert a loop inside the loop. The example of doing so is given below.